Fixed PORT and message error in user controllers

+++ backend/controllers/user.js
@@ -52,11 +52,17 @@ export const updateUser = async (req, res) => {
 
 export const deleteUser = async (req, res) => {
     const {id} = req.params;
+
+    if(!mongoose.Types.ObjectId.isValid(id))
+    {
+        return res.status(404).json({success: false, message: "User not found - invalid id"});
+    }
+
     try{
         await User.findByIdAndDelete(id);
         res.status(200).json({success: true, message: 'User deleted successfully'});
     }catch (error) {
         console.log("Error in Delete user:", error.message);
-        res.status(404).json({success: false, message:"User not found"});
+        res.status(500).json({success: false, message:"Server error"});
     }
 }
